Replace pom-documentation.xml by maven profile
Unfortunately I was unable to properly test this. My github user isn't allowed
to download the springboot-swagger-3 plugin from the maven repository. The
version 1.2.4 does not have a (pushed) tag in the git repository, so I used
commit 39069b9 (which than again required springboot-swagger3-annotations
0.0.5 which didn't have a tag and the github releases featured version numbers
not reflected in the code, e.g. 0.3.1, so commit ae97c7f was used).